Split workspace execution parity gates
